refactor: compnent 파일 분류
